requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain_google_genai
faiss-cpu
PyPDF2
langchain_community